Add files via upload
Learned about Figure in html along with figcaption.
Worked heavily to understand how the transform/translate/position of the caption text work both before and after hover. It was a struggle and I need more practice on that aspect.
Learned about shape-outside and how that does not work on Microsoft Edge :'(
Learned more about filter by using blur and brightness.