Media Queries starting place.
Learned how to create a media query manager using Sass. Creating a mixin based the different widths on the screen. Learned that the fast way to incorpoate this code is to use a $breakpoint to define the type of phone being used. Additionally found out that em is considered the best form of measurement in writing media queries.